ref: move CsvGenerator and EncryptedResponseCsvGenerator to own dir
in preparation for colocation of tests
