+++ frontend/src/features/admin-form/responses/ResponsesPage/common/utils/index.ts
@@ -1,1 +1,1 @@
-export * from './CsvGenerator.class'
+export * from './CsvGenerator'
