+++ frontend/src/features/admin-form/responses/ResponsesPage/storage/utils/EncryptedResponseCsvGenerator/EncryptedResponseCsvGenerator.ts
@@ -4,11 +4,10 @@ import type { Dictionary } from 'lodash'
 import { keyBy } from 'lodash'
 import type { Merge } from 'type-fest'
 
-import { CsvGenerator } from '../../common/utils'
-import type { DecryptedSubmissionData } from '../types'
-
-import type { Response } from './csv-response-classes'
-import { getDecryptedResponseInstance } from './getDecryptedResponseInstance'
+import { CsvGenerator } from '../../../common/utils'
+import type { DecryptedSubmissionData } from '../../types'
+import type { Response } from '../csv-response-classes'
+import { getDecryptedResponseInstance } from '../getDecryptedResponseInstance'
 
 type UnprocessedRecord = Merge<
   DecryptedSubmissionData,
