+++ frontend/src/features/admin-form/responses/ResponsesPage/storage/utils/EncryptedResponseCsvGenerator/index.ts
@@ -1,0 +1,1 @@
+export { EncryptedResponseCsvGenerator } from './EncryptedResponseCsvGenerator'
